requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.8.1
    # via flask-migrate
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
babel==2.10.3
    # via
    #   flask-babel
    #   flask-babelex
barcode==1.0.2
    # via -r requirements.in
bidict==0.22.0
    # via python-socketio
blinker==1.5
    # via
    #   flask-mail
    #   flask-principal
bokeh==2.4.3
    # via -r requirements.in
click==8.1.3
    # via
    #   -r requirements.in
    #   flask
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via -r requirements.in
dnspython==2.2.1
    # via email-validator
dominate==2.7.0
    # via flask-bootstrap
email-validator==1.3.0
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-admin
    #   flask-alchemydumps
    #   flask-babel
    #   flask-babelex
    #   flask-bootstrap
    #   flask-cors
    #   flask-log-request-id
    #   flask-login
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-principal
    #   flask-security
    #   flask-socketio
    #   flask-sqlalchemy
    #   flask-wtf
    #   pytest-flask
flask-admin==1.6.0
    # via -r requirements.in
flask-alchemydumps==0.0.13
    # via -r requirements.in
flask-babel==2.0.0
    # via -r requirements.in
flask-babelex==0.9.4
    # via
    #   -r requirements.in
    #   flask-security
flask-bootstrap==3.3.7.1
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask-log-request-id==0.10.1
    # via -r requirements.in
flask-login==0.6.2
    # via
    #   -r requirements.in
    #   flask-security
flask-mail==0.9.1
    # via
    #   -r requirements.in
    #   flask-security
flask-marshmallow==0.14.0
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-principal==0.4.0
    # via flask-security
flask-security==3.0.0
    # via -r requirements.in
flask-socketio==5.3.1
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-alchemydumps
    #   flask-migrate
flask-wtf==1.0.1
    # via
    #   -r requirements.in
    #   flask-security
fonttools==4.37.4
    # via matplotlib
fpdf==1.7.2
    # via -r requirements.in
greenlet==1.1.3
    # via sqlalchemy
idna==3.4
    # via email-validator
imageio==2.22.1
    # via -r requirements.in
importlib-metadata==5.0.0
    # via
    #   alembic
    #   click
    #   flask
    #   jsonschema
    #   mako
    #   markdown
    #   pluggy
    #   pytest
    #   sqlalchemy
    #   sqlalchemy-utils
importlib-resources==5.9.0
    # via
    #   alembic
    #   jsonschema
imutils==0.5.4
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-security
    #   flask-wtf
jinja2==3.1.2
    # via
    #   bokeh
    #   flask
    #   flask-babel
    #   flask-babelex
jsonschema==4.16.0
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
levenshtein==0.20.5
    # via python-levenshtein
lxml==4.9.2
    # via -r requirements.in
mako==1.2.3
    # via alembic
markdown==3.4.1
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==3.18.0
    # via
    #   -r requirements.in
    #   flask-marshmallow
    #   marshmallow-jsonapi
    #   marshmallow-sqlalchemy
marshmallow-jsonapi==0.24.0
    # via -r requirements.in
marshmallow-sqlalchemy==0.28.1
    # via -r requirements.in
matplotlib==3.5.3
    # via -r requirements.in
mirakuru==2.5.1
    # via pytest-postgresql
numpy==1.21.6
    # via
    #   -r requirements.in
    #   bokeh
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pandas
    #   tifffile
opencv-python==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via
    #   bokeh
    #   marshmallow
    #   marshmallow-sqlalchemy
    #   matplotlib
    #   pytesseract
    #   pytest
pandas==1.3.5
    # via -r requirements.in
passlib==1.7.4
    # via flask-security
pdf2image==1.16.0
    # via -r requirements.in
pdfkit==1.0.0
    # via -r requirements.in
pillow==9.2.0
    # via
    #   -r requirements.in
    #   bokeh
    #   imageio
    #   matplotlib
    #   pdf2image
    #   pytesseract
    #   reportlab
pkgutil-resolve-name==1.3.10
    # via jsonschema
pluggy==1.0.0
    # via pytest
port-for==0.7.0
    # via pytest-postgresql
protobuf==4.21.7
    # via -r requirements.in
psutil==5.9.5
    # via mirakuru
psycopg2-binary==2.9.4
    # via -r requirements.in
py==1.11.0
    # via pytest
pydantic==1.10.2
    # via
    #   -r requirements.in
    #   pydantic-loader
pydantic-loader[toml]==0.9.1
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   matplotlib
    #   packaging
pypdf2==2.11.1
    # via -r requirements.in
pyrsistent==0.18.1
    # via jsonschema
pytesseract==0.3.10
    # via -r requirements.in
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-disable-plugin
    #   pytest-flask
    #   pytest-postgresql
pytest-disable-plugin==0.1.2
    # via -r requirements.in
pytest-flask==1.2.0
    # via -r requirements.in
pytest-postgresql==4.1.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
python-decouple==3.6
    # via flask-alchemydumps
python-dotenv==0.21.0
    # via -r requirements.in
python-engineio==4.3.4
    # via python-socketio
python-levenshtein==0.20.5
    # via barcode
python-magic==0.4.27
    # via -r requirements.in
python-socketio==5.7.1
    # via
    #   -r requirements.in
    #   flask-socketio
pytz==2022.4
    # via
    #   babel
    #   flask-babel
    #   pandas
pyyaml==6.0
    # via bokeh
pyzbar==0.1.9
    # via -r requirements.in
rapidfuzz==2.11.1
    # via levenshtein
reportlab==3.6.11
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-cors
    #   flask-marshmallow
    #   python-dateutil
sortedcontainers==2.4.0
    # via -r requirements.in
speaklater==1.3
    # via flask-babelex
sqlalchemy==1.4.41
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-querybuilder
    #   sqlalchemy-utils
sqlalchemy-querybuilder==0.1.5
    # via -r requirements.in
sqlalchemy-utils==0.38.3
    # via -r requirements.in
tifffile==2021.11.2
    # via -r requirements.in
toml==0.10.2
    # via
    #   -r requirements.in
    #   pydantic-loader
tomli==2.0.1
    # via pytest
tornado==6.2
    # via bokeh
typing-extensions==4.3.0
    # via
    #   bokeh
    #   importlib-metadata
    #   jsonschema
    #   kiwisolver
    #   pydantic
    #   pypdf2
ua-parser==0.16.1
    # via
    #   -r requirements.in
    #   user-agents
user-agents==2.2.0
    # via -r requirements.in
visitor==0.1.3
    # via flask-bootstrap
werkzeug==2.2.2
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
    #   pytest-flask
wtforms==3.0.1
    # via
    #   flask-admin
    #   flask-wtf
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
